Update.
	* sysdeps/unix/sysv/linux/powerpc/rt-sysdep.c: New file.
	* sysdeps/unix/sysv/linux/powerpc/Makefile (librt-routines): Add
	rt-sysdep.

	* rt/tst-mqueue5.c (mqsend): Don't inline.
	(mqrecv): Likewise.

+++ nptl/sysdeps/unix/sysv/linux/sparc/bits/semaphore.h
@@ -34,7 +34,7 @@
 #define SEM_FAILED      ((sem_t *) 0)
 
 /* Maximum value the semaphore can have.  */
-#define SEM_VALUE_MAX   ((int) (2147483647)
+#define SEM_VALUE_MAX   (2147483647)
 
 
 typedef union
